requirements.txt
# requirements.txt

# Core FastAPI and ASGI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0
sqlite-utils>=3.35.2

# AI/LLM Integration
openai>=1.6.1
tiktoken>=0.5.2

# Authentication & Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Configuration & Environment
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Data Processing & Analysis
pandas>=2.1.4
numpy>=1.25.2
python-dateutil>=2.8.2

# HTTP & Requests
httpx>=0.25.2
aiohttp>=3.9.1
requests>=2.31.0

# Template Engine
jinja2>=3.1.2

# File Processing
python-magic>=0.4.27
openpyxl>=3.1.2
python-docx>=1.1.0

# Utilities
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# Development & Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Production/Deployment
gunicorn>=21.2.0
python-json-logger>=2.0.7

# Monitoring & Logging
structlog>=23.2.0
loguru>=0.7.2

# Background Tasks (optional)
celery>=5.3.4
redis>=5.0.1

# Data Visualization (for admin analytics)
plotly>=5.17.0
matplotlib>=3.8.2

# Natural Language Processing (for enhanced representations)
spacy>=3.7.2
nltk>=3.8.1

# Export capabilities
reportlab>=4.0.7
weasyprint>=60.2

# Memory optimization
psutil>=5.9.6

# Async utilities
asyncio-mqtt>=0.14.1
